Add initial implementation of updraft dev server with reloading
- Adapted from werkzeug's "simple server"
- Only "simple" reloader is used. Werkzeug's Watchdog-based
reloader is not implemented
- Apache license for the repo as a whole, with BSD copyright/licence
notice and references for files that are adapted from Werkzeug. Not
sure if this part is done correctly.
- Known flakiness in test that verifies reloader behavior.
Papered over for the time being with flaky, but I'm not sure
why it's happening. Since the reloader comes from Werkzeug
(albeit in modified form), which is pretty stable, it seems
like it might be a problem with the test setup, but I never
pinned down the cause.